test: TDD Qualifier 테스트 작성

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -2,13 +2,17 @@ package woowacourse.shopping.data
 
 import woowacourse.shopping.data.mapper.toDomain
 import woowacourse.shopping.data.mapper.toEntity
+import woowacourse.shopping.di.annotation.Binds
+import woowacourse.shopping.di.annotation.Provides
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.model.DatabaseIdentifier
 import woowacourse.shopping.model.Identifier
 import woowacourse.shopping.model.Product
 
+@Provides
+@Room
 class DefaultCartRepository(
-    private val cartProductDao: CartProductDao
+    @Binds @Room private val cartProductDao: CartProductDao
 ) : CartRepository {
     override suspend fun addCartProduct(product: Product) {
         cartProductDao.insert(product.toEntity())
@@ -21,7 +25,6 @@ class DefaultCartRepository(
     }
 
     override suspend fun deleteCartProduct(id: Identifier<*>) {
-        if (id is DatabaseIdentifier)
-            cartProductDao.delete(id.value)
+        if (id is DatabaseIdentifier) cartProductDao.delete(id.value)
     }
 }
